+++ Plugins/MeshWidget/Binaries/Win64/UE4Editor.modules
@@ -1,0 +1,7 @@
+{
+	"BuildId": "4212847",
+	"Modules": 
+	{
+		"MeshWidget": "UE4Editor-MeshWidget.dll"
+	}
+}
