--- Plugins/Runtime/MeshWidget/Binaries/Win64/UE4Editor-Win64-DebugGame.modules
@@ -1,8 +1,0 @@
-{
-	"Changelist" : 3106830,
-	"BuildId" : "cb1ef911-c361-4ca8-9f3c-48e00da0d839",
-	"Modules" : 
-	{
-		"MeshWidget" : "UE4Editor-MeshWidget-Win64-DebugGame.dll"
-	}
-}
