--- Plugins/Runtime/MeshWidget/Binaries/Win64/UE4Editor.modules
@@ -1,8 +1,0 @@
-{
-	"Changelist" : 3142249,
-	"BuildId" : "7fd9829c-8021-4b62-b69f-09e2c3cd6230",
-	"Modules" : 
-	{
-		"MeshWidget" : "UE4Editor-MeshWidget-8598.dll"
-	}
-}
