--- Plugins/Runtime/MeshWidget/Source/MeshWidget/Private/MeshWidgetInteractionComponent.cpp
@@ -1,530 +1,0 @@
-// Copyright 1998-2015 Epic Games, Inc. All Rights Reserved.
-
-#include "MeshWidgetPrivatePCH.h"
-
-#include "WidgetLayoutLibrary.h"
-#include "MeshWidgetInteractionComponent.h"
-
-#include "Kismet/KismetSystemLibrary.h"
-
-#define LOCTEXT_NAMESPACE "WidgetInteraction"
-
-UMeshWidgetInteractionComponent::UMeshWidgetInteractionComponent(const FObjectInitializer& ObjectInitializer)
-	: Super(ObjectInitializer)
-	, VirtualUserIndex(0)
-	, PointerIndex(0)
-	, InteractionDistance(500)
-	, InteractionSource(EWidgetInteractionSource::World)
-	, bEnableHitTesting(true)
-	, bShowDebug(false)
-	, DebugColor(FLinearColor::Red)
-{
-	PrimaryComponentTick.bCanEverTick = true;
-
-#if WITH_EDITORONLY_DATA
-	ArrowComponent = ObjectInitializer.CreateEditorOnlyDefaultSubobject<UArrowComponent>(this, TEXT("ArrowComponent0"));
-
-	if ( ArrowComponent && !IsTemplate() )
-	{
-		ArrowComponent->ArrowColor = DebugColor.ToFColor(true);
-		ArrowComponent->AttachToComponent(this, FAttachmentTransformRules(EAttachmentRule::KeepRelative, false));
-	}
-#endif
-}
-
-void UMeshWidgetInteractionComponent::OnComponentCreated()
-{
-#if WITH_EDITORONLY_DATA
-	if ( ArrowComponent )
-	{
-		ArrowComponent->ArrowColor = DebugColor.ToFColor(true);
-		ArrowComponent->SetVisibility(bEnableHitTesting);
-	}
-#endif
-}
-
-void UMeshWidgetInteractionComponent::BeginPlay()
-{
-	Super::BeginPlay();
-
-	if ( FSlateApplication::IsInitialized() )
-	{
-		if ( !VirtualUser.IsValid() )
-		{
-			VirtualUser = FSlateApplication::Get().FindOrCreateVirtualUser(VirtualUserIndex);
-		}
-	}
-}
-
-void UMeshWidgetInteractionComponent::EndPlay(const EEndPlayReason::Type EndPlayReason)
-{
-	Super::EndPlay(EndPlayReason);
-
-	if ( FSlateApplication::IsInitialized() )
-	{
-		if ( VirtualUser.IsValid() )
-		{
-			FSlateApplication::Get().UnregisterUser(VirtualUser->GetUserIndex());
-			VirtualUser.Reset();
-		}
-	}
-}
-
-void UMeshWidgetInteractionComponent::TickComponent(float DeltaTime, ELevelTick TickType, FActorComponentTickFunction* ThisTickFunction)
-{
-	Super::TickComponent(DeltaTime, TickType, ThisTickFunction);
-
-	SimulatePointerMovement();
-}
-
-bool UMeshWidgetInteractionComponent::CanSendInput()
-{
-	return FSlateApplication::IsInitialized() && VirtualUser.IsValid();
-}
-
-void UMeshWidgetInteractionComponent::SetCustomHitResult(const FHitResult& HitResult)
-{
-	CustomHitResult = HitResult;
-}
-
-bool UMeshWidgetInteractionComponent::PerformTrace(FHitResult& HitResult)
-{
-	FCollisionQueryParams Params = FCollisionQueryParams::DefaultQueryParam;
-	Params.bTraceComplex = true;
-	Params.bReturnFaceIndex = true;
-
-	switch( InteractionSource )
-	{
-		case EWidgetInteractionSource::World:
-		{
-			const FVector WorldLocation = GetComponentLocation();
-			const FTransform WorldTransform = GetComponentTransform();
-			const FVector Direction = WorldTransform.GetUnitAxis(EAxis::X);
-
-			TArray<UPrimitiveComponent*> PrimitiveChildren;
-			GetRelatedComponentsToIgnoreInAutomaticHitTesting(PrimitiveChildren);
-
-			Params.AddIgnoredComponents(PrimitiveChildren);
-
-			FCollisionObjectQueryParams Everything(FCollisionObjectQueryParams::AllObjects);
-			return GetWorld()->LineTraceSingleByObjectType(HitResult, WorldLocation, WorldLocation + ( Direction * InteractionDistance ), Everything, Params);
-		}
-		case EWidgetInteractionSource::Mouse:
-		case EWidgetInteractionSource::CenterScreen:
-		{
-			TArray<UPrimitiveComponent*> PrimitiveChildren;
-			GetRelatedComponentsToIgnoreInAutomaticHitTesting(PrimitiveChildren);
-
-			Params.AddIgnoredComponents(PrimitiveChildren);
-
-			APlayerController* PlayerController = GetWorld()->GetFirstPlayerController();
-
-			ULocalPlayer* LocalPlayer = PlayerController->GetLocalPlayer();
-			bool bHit = false;
-			if ( LocalPlayer && LocalPlayer->ViewportClient )
-			{
-				if ( InteractionSource == EWidgetInteractionSource::Mouse )
-				{
-					FVector2D MousePosition;
-					if ( LocalPlayer->ViewportClient->GetMousePosition(MousePosition) )
-					{
-						bHit = PlayerController->GetHitResultAtScreenPosition(MousePosition, ECC_Visibility, Params, HitResult);
-					}
-				}
-				else if ( InteractionSource == EWidgetInteractionSource::CenterScreen )
-				{
-					FVector2D ViewportSize;
-					LocalPlayer->ViewportClient->GetViewportSize(ViewportSize);
-
-					bHit = PlayerController->GetHitResultAtScreenPosition(ViewportSize * 0.5f, ECC_Visibility, Params, HitResult);
-				}
-
-				// Don't allow infinite distance hit testing.
-				if ( bHit )
-				{
-					if ( HitResult.Distance > InteractionDistance )
-					{
-						HitResult = FHitResult();
-						bHit = false;
-					}
-				}
-			}
-			
-			return bHit;
-		}
-		case EWidgetInteractionSource::Custom:
-		{
-			HitResult = CustomHitResult;
-			return HitResult.bBlockingHit;
-		}
-	}
-
-	return false;
-}
-
-void UMeshWidgetInteractionComponent::GetRelatedComponentsToIgnoreInAutomaticHitTesting(TArray<UPrimitiveComponent*>& IgnorePrimitives)
-{
-	TArray<USceneComponent*> SceneChildren;
-	if ( AActor* Owner = GetOwner() )
-	{
-		if ( USceneComponent* Root = Owner->GetRootComponent() )
-		{
-			Root = Root->GetAttachmentRoot();
-			Root->GetChildrenComponents(true, SceneChildren);
-			SceneChildren.Add(Root);
-		}
-	}
-
-	for ( USceneComponent* SceneComponent : SceneChildren )
-	{
-		if ( UPrimitiveComponent* PrimtiveComponet = Cast<UPrimitiveComponent>(SceneComponent) )
-		{
-			// Don't ignore widget components that are siblings.
-			if ( SceneComponent->IsA<UWidgetComponent>() || SceneComponent->IsA<UMeshWidgetComponent>())
-			{
-				continue;
-			}
-
-			IgnorePrimitives.Add(PrimtiveComponet);
-		}
-	}
-}
-
-void UMeshWidgetInteractionComponent::SimulatePointerMovement()
-{
-	bIsHoveredWidgetInteractable = false;
-	bIsHoveredWidgetFocusable = false;
-	bIsHoveredWidgetHitTestVisible = false;
-
-	if ( !bEnableHitTesting )
-	{
-		return;
-	}
-
-	if ( !CanSendInput() )
-	{
-		return;
-	}
-	
-	LocalHitLocation = FVector2D(0, 0);
-	FWidgetPath WidgetPathUnderFinger;
-	
-	const bool bHit = PerformTrace(LastHitResult);
-
-	UWidgetComponent* OldHoveredWidget = HoveredWidgetComponent;
-	UMeshWidgetComponent* OldHoveredMeshWidget = HoveredMeshWidgetComponent;
-	
-	HoveredWidgetComponent = nullptr;
-	HoveredMeshWidgetComponent = nullptr;
-
-	if (bHit)
-	{
-		HoveredWidgetComponent = Cast<UWidgetComponent>(LastHitResult.GetComponent());
-		if ( HoveredWidgetComponent )
-		{
-			HoveredWidgetComponent->GetLocalHitLocation(LastHitResult.ImpactPoint, LocalHitLocation );
-			WidgetPathUnderFinger = FWidgetPath(HoveredWidgetComponent->GetHitWidgetPath(LastHitResult.ImpactPoint, /*bIgnoreEnabledStatus*/ false));
-		}
-		HoveredMeshWidgetComponent = Cast<UMeshWidgetComponent>(LastHitResult.GetComponent());
-		if ( HoveredMeshWidgetComponent )
-		{
-			LocalHitLocation = HoveredMeshWidgetComponent->GetLocalHitLocation(LastHitResult);
-			WidgetPathUnderFinger = FWidgetPath(HoveredMeshWidgetComponent->GetHitWidgetPath(LastHitResult, /*bIgnoreEnabledStatus*/ false));
-		}
-	}
-
-	if ( bShowDebug )
-	{
-		if ( HoveredWidgetComponent )
-		{
-			UKismetSystemLibrary::DrawDebugSphere(this, LastHitResult.ImpactPoint, 2.5f, 12, DebugColor, 0, 2);
-		}
-		if ( HoveredMeshWidgetComponent )
-		{
-			UKismetSystemLibrary::DrawDebugSphere(this, LastHitResult.ImpactPoint, 2.5f, 12, DebugColor, 0, 2);
-		}
-
-		if ( InteractionSource == EWidgetInteractionSource::World || InteractionSource == EWidgetInteractionSource::Custom )
-		{
-			if ( HoveredWidgetComponent || HoveredMeshWidgetComponent )
-			{
-				UKismetSystemLibrary::DrawDebugLine(this, LastHitResult.TraceStart, LastHitResult.ImpactPoint, DebugColor, 0, 1);
-			}
-			else
-			{
-				UKismetSystemLibrary::DrawDebugLine(this, LastHitResult.TraceStart, LastHitResult.TraceEnd, DebugColor, 0, 1);
-			}
-		}
-	}
-	
-	FPointerEvent PointerEvent(
-		VirtualUser->GetUserIndex(),
-		PointerIndex,
-		LocalHitLocation,
-		LastLocalHitLocation,
-		PressedKeys,
-		FKey(),
-		0.0f,
-		ModifierKeys);
-	
-	if (WidgetPathUnderFinger.IsValid())
-	{
-		check(HoveredWidgetComponent || HoveredMeshWidgetComponent);
-		LastWigetPath = WidgetPathUnderFinger;
-		
-		FSlateApplication::Get().RoutePointerMoveEvent(WidgetPathUnderFinger, PointerEvent, false);
-	}
-	else
-	{
-		FWidgetPath EmptyWidgetPath;
-		FSlateApplication::Get().RoutePointerMoveEvent(EmptyWidgetPath, PointerEvent, false);
-
-		LastWigetPath = FWeakWidgetPath();
-	}
-
-	if ( HoveredWidgetComponent )
-	{
-		HoveredWidgetComponent->RequestRedraw();
-	}
-	if ( HoveredMeshWidgetComponent )
-	{
-		HoveredMeshWidgetComponent->RequestRedraw();
-	}
-
-	LastLocalHitLocation = LocalHitLocation;
-
-	if ( WidgetPathUnderFinger.IsValid() )
-	{
-		const FArrangedChildren::FArrangedWidgetArray& AllArrangedWidgets = WidgetPathUnderFinger.Widgets.GetInternalArray();
-		for ( const FArrangedWidget& ArrangedWidget : AllArrangedWidgets )
-		{
-			const TSharedRef<SWidget>& Widget = ArrangedWidget.Widget;
-			if ( Widget->IsInteractable() )
-			{
-				bIsHoveredWidgetInteractable = true;
-			}
-
-			if ( Widget->SupportsKeyboardFocus() )
-			{
-				bIsHoveredWidgetFocusable = true;
-			}
-
-			if ( Widget->GetVisibility().IsHitTestVisible() )
-			{
-				bIsHoveredWidgetHitTestVisible = true;
-			}
-		}
-	}
-
-	if ( HoveredWidgetComponent != OldHoveredWidget )
-	{
-		if ( OldHoveredWidget )
-		{
-			OldHoveredWidget->RequestRedraw();
-		}
-
-		OnHoveredWidgetChanged.Broadcast(HoveredWidgetComponent, OldHoveredWidget);
-	}
-	if ( HoveredMeshWidgetComponent != OldHoveredMeshWidget )
-	{
-		if ( OldHoveredMeshWidget )
-		{
-			OldHoveredMeshWidget->RequestRedraw();
-		}
-
-		OnHoveredMeshWidgetChanged.Broadcast(HoveredMeshWidgetComponent, OldHoveredMeshWidget);
-	}
-}
-
-void UMeshWidgetInteractionComponent::PressPointerKey(FKey Key)
-{
-	if ( !CanSendInput() )
-	{
-		return;
-	}
-
-	if (PressedKeys.Contains(Key))
-	{
-		return;
-	}
-	
-	PressedKeys.Add(Key);
-	
-	FWidgetPath WidgetPathUnderFinger = LastWigetPath.ToWidgetPath();
-		
-	FPointerEvent PointerEvent(
-		VirtualUser->GetUserIndex(),
-		PointerIndex,
-		LocalHitLocation,
-		LastLocalHitLocation,
-		PressedKeys,
-		Key,
-		0.0f,
-		ModifierKeys);
-		
-	FReply Reply = FSlateApplication::Get().RoutePointerDownEvent(WidgetPathUnderFinger, PointerEvent);
-	
-	// @TODO Something about double click, expose directly, or automatically do it if key press happens within
-	// the double click timeframe?
-	//Reply = FSlateApplication::Get().RoutePointerDoubleClickEvent( WidgetPathUnderFinger, PointerEvent );
-}
-
-void UMeshWidgetInteractionComponent::ReleasePointerKey(FKey Key)
-{
-	if ( !CanSendInput() )
-	{
-		return;
-	}
-
-	if (!PressedKeys.Contains(Key))
-	{
-		return;
-	}
-	
-	PressedKeys.Remove(Key);
-	
-	FWidgetPath WidgetPathUnderFinger = LastWigetPath.ToWidgetPath();
-		
-	FPointerEvent PointerEvent(
-		VirtualUser->GetUserIndex(),
-		PointerIndex,
-		LocalHitLocation,
-		LastLocalHitLocation,
-		PressedKeys,
-		Key,
-		0.0f,
-		ModifierKeys);
-		
-	FReply Reply = FSlateApplication::Get().RoutePointerUpEvent(WidgetPathUnderFinger, PointerEvent);
-}
-
-bool UMeshWidgetInteractionComponent::PressKey(FKey Key, bool bRepeat)
-{
-	if ( !CanSendInput() )
-	{
-		return false;
-	}
-
-	const uint32* KeyCodePtr;
-	const uint32* CharCodePtr;
-	FInputKeyManager::Get().GetCodesFromKey(Key, KeyCodePtr, CharCodePtr);
-
-	uint32 KeyCode = KeyCodePtr ? *KeyCodePtr : 0;
-	uint32 CharCode = CharCodePtr ? *CharCodePtr : 0;
-
-	FKeyEvent KeyEvent(Key, ModifierKeys, VirtualUser->GetUserIndex(), bRepeat, KeyCode, CharCode);
-	bool DownResult = FSlateApplication::Get().ProcessKeyDownEvent(KeyEvent);
-
-	if ( CharCodePtr )
-	{
-		FCharacterEvent CharacterEvent(CharCode, ModifierKeys, VirtualUser->GetUserIndex(), bRepeat);
-		return FSlateApplication::Get().ProcessKeyCharEvent(CharacterEvent);
-	}
-
-	return DownResult;
-}
-
-bool UMeshWidgetInteractionComponent::ReleaseKey(FKey Key)
-{
-	if ( !CanSendInput() )
-	{
-		return false;
-	}
-
-	const uint32* KeyCodePtr;
-	const uint32* CharCodePtr;
-	FInputKeyManager::Get().GetCodesFromKey(Key, KeyCodePtr, CharCodePtr);
-
-	uint32 KeyCode = KeyCodePtr ? *KeyCodePtr : 0;
-	uint32 CharCode = CharCodePtr ? *CharCodePtr : 0;
-
-	FKeyEvent KeyEvent(Key, ModifierKeys, VirtualUser->GetUserIndex(), false, KeyCode, CharCode);
-	return FSlateApplication::Get().ProcessKeyUpEvent(KeyEvent);
-}
-
-bool UMeshWidgetInteractionComponent::PressAndReleaseKey(FKey Key)
-{
-	const bool PressResult = PressKey(Key, false);
-	const bool ReleaseResult = ReleaseKey(Key);
-
-	return PressResult || ReleaseResult;
-}
-
-bool UMeshWidgetInteractionComponent::SendKeyChar(FString Characters, bool bRepeat)
-{
-	if ( !CanSendInput() )
-	{
-		return false;
-	}
-
-	bool bProcessResult = false;
-
-	for ( int32 CharIndex = 0; CharIndex < Characters.Len(); CharIndex++ )
-	{
-		TCHAR CharKey = Characters[CharIndex];
-
-		FCharacterEvent CharacterEvent(CharKey, ModifierKeys, VirtualUser->GetUserIndex(), bRepeat);
-		bProcessResult |= FSlateApplication::Get().ProcessKeyCharEvent(CharacterEvent);
-	}
-
-	return bProcessResult;
-}
-
-void UMeshWidgetInteractionComponent::ScrollWheel(float ScrollDelta)
-{
-	if ( !CanSendInput() )
-	{
-		return;
-	}
-
-	FWidgetPath WidgetPathUnderFinger = LastWigetPath.ToWidgetPath();
-	
-	FPointerEvent MouseWheelEvent(
-		VirtualUser->GetUserIndex(),
-		PointerIndex,
-		LocalHitLocation,
-		LastLocalHitLocation,
-		PressedKeys,
-		EKeys::MouseWheelAxis,
-		ScrollDelta,
-		ModifierKeys);
-
-	FSlateApplication::Get().RouteMouseWheelOrGestureEvent(WidgetPathUnderFinger, MouseWheelEvent, nullptr);
-}
-
-UWidgetComponent* UMeshWidgetInteractionComponent::GetHoveredWidgetComponent() const
-{
-	return HoveredWidgetComponent;
-}
-
-UMeshWidgetComponent* UMeshWidgetInteractionComponent::GetHoveredMeshWidgetComponent() const
-{
-	return HoveredMeshWidgetComponent;
-}
-
-bool UMeshWidgetInteractionComponent::IsOverInteractableWidget() const
-{
-	return bIsHoveredWidgetInteractable;
-}
-
-bool UMeshWidgetInteractionComponent::IsOverFocusableWidget() const
-{
-	return bIsHoveredWidgetFocusable;
-}
-
-bool UMeshWidgetInteractionComponent::IsOverHitTestVisibleWidget() const
-{
-	return bIsHoveredWidgetHitTestVisible;
-}
-
-const FWeakWidgetPath& UMeshWidgetInteractionComponent::GetHoveredWidgetPath() const
-{
-	return LastWigetPath;
-}
-
-const FHitResult& UMeshWidgetInteractionComponent::GetLastHitResult() const
-{
-	return LastHitResult;
-}
-
-#undef LOCTEXT_NAMESPACE
