--- Plugins/Runtime/MeshWidget/Source/MeshWidget/Private/MeshWidgetPrivatePCH.h
@@ -1,8 +1,0 @@
-// Copyright 1998-2016 Epic Games, Inc. All Rights Reserved.
-
-#include "IMeshWidget.h"
-
-// You should place include statements to your module's private header files here.  You only need to
-// add includes for headers that are used in most of your module's source files though.
-
-#include "UMG.h"
