+++ Source/MeshWidgets/MeshWidgetsGame.cpp
@@ -1,8 +1,4 @@
 // Fill out your copyright notice in the Description page of Project Settings.
 
-#include "MeshWidgets.h"
+
 #include "MeshWidgetsGame.h"
-
-
-
-
