+++ Source/MeshWidgetsEditor.Target.cs
@@ -1,25 +1,13 @@
-// Fill out your copyright notice in the Description page of Project Settings.
 
 using UnrealBuildTool;
 using System.Collections.Generic;
 
 public class MeshWidgetsEditorTarget : TargetRules
 {
-	public MeshWidgetsEditorTarget(TargetInfo Target)
+	public MeshWidgetsEditorTarget(TargetInfo Target) : base(Target)
 	{
 		Type = TargetType.Editor;
-	}
 
-	//
-	// TargetRules interface.
-	//
-
-	public override void SetupBinaries(
-		TargetInfo Target,
-		ref List<UEBuildBinaryConfiguration> OutBuildBinaryConfigurations,
-		ref List<string> OutExtraModuleNames
-		)
-	{
-		OutExtraModuleNames.AddRange( new string[] { "MeshWidgets" } );
+		ExtraModuleNames.AddRange( new string[] { "MeshWidgets" } );
 	}
 }
